requirement.txt
Pillow
pywin32>=227
pypiwin32>=223
